Compare the ACM's compression ratio without flooring it first
In smart-output mode the codec derives an output sample rate from the
compression ratio and then refuses any conversion whose destination does
not match it, so this ratio decides which conversions are possible at
all.  Every operand of the expression is an integer, so the division
truncated before the comparison and a ratio of 13.7 was treated as 13.

Three source and destination pairs are affected, all of them at that
ratio: 24 kHz stereo at 56 kbps, 48 kHz mono at 56 kbps and 48 kHz
stereo at 112 kbps.  Each of them now downsamples as intended instead of
keeping the source rate.

+++ lame/ACM/ACMStream.cpp
@@ -385,7 +385,11 @@ unsigned int ACMStream::GetOutputSampleRate(int samples_per_sec, int bitrate, in
   
         /// \todo pass through the same LAME routine
 	unsigned int OutputFrequency;
-	double compression_ratio = double(samples_per_sec * 16 * channels / (bitrate * 8));
+	// Both operands are ints, so the division is deliberately written on
+	// doubles: done in integer arithmetic it floors the ratio before the
+	// comparison below, which puts a ratio of 13.7 on the wrong side of the
+	// boundary and offers a rate the caller then refuses to convert to.
+	double compression_ratio = double(samples_per_sec * 16 * channels) / double(bitrate * 8);
 	// map2MP3Frequency() takes a frequency in Hz as an int, so the fractional
 	// part of both expressions has always been discarded here.  The casts
 	// write that down; they do not change which rate comes back.
